+++ docs/contents/changelog.rst
@@ -22,6 +22,12 @@ Fixed
 
 - Fixed catalog outline title resolve issue (`PR #10 <https://github.com/useblocks/libpdf/pull/10>`_)
 
+Added
+~~~~~
+
+- Added flag ``--no_annotations`` to exclude annotation extraction from the catalog to speed up extraction
+  (`PR #15 <https://github.com/useblocks/libpdf/pull/15>`_)
+
 __ https://github.com/useblocks/libpdf/releases/tag/v0.0.1
 
 `0.0.1`__ - 2020-06-30